requirements.docker.txt
# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0

# Authentication & Security
authlib>=1.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP Client & Utilities
httpx>=0.25.0
email-validator>=2.0.0

# AI & LangGraph
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
openai>=1.3.0

# Media Processing
pillow>=10.0.0
python-ffmpeg>=2.0.0

# Real-time & Caching
redis>=4.6.0
websockets>=11.0.0

# File Storage & Upload
aiofiles>=23.0.0

# Web Scraping & Data Extraction
requests>=2.31.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
selenium>=4.15.0
webdriver-manager>=4.0.0
instaloader>=4.9.0
lxml-html-clean==0.2.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Video Processing
pytubefix>=6.0.0

# Vector Database
pymilvus>=2.3.0

# Machine Learning & NLP
huggingface-hub>=0.17.0

# Database Drivers
psycopg2-binary>=2.9.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
